+++ src/main/java/com/webank/webase/front/base/enums/EventTypes.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright 2014-2019 the original author or authors.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webank.webase.front.base.enums;
+
+public enum EventTypes {
+
+	/**
+	 * block notify type
+	 */
+	BLOCK_NOTIFY(1),
+
+	/**
+	 * contract's event log push type
+	 */
+	EVENT_LOG_PUSH(2),
+
+	/**
+	 * others ex: sdk error message
+	 */
+	OTHERS(3);
+
+	private int value;
+
+	EventTypes(Integer eventType) {
+		this.value = eventType;
+	}
+
+	public int getValue() {
+		return this.value;
+	}
+}
