+++ src/main/java/com/webank/webase/front/event/RabbitMQPublisher.java
