+++ src/main/java/com/webank/webase/front/event/entity/message/BlockPushMessage.java
