+++ src/main/java/com/webank/webase/front/event/entity/message/MQObject.java
